move and rename package with ros communication nodes -> seerep_ros_comm
