+++ seerep_ros/seerep_ros_comm/CMakeLists.txt
@@ -1,7 +1,7 @@
 #catkin_lint: ignore critical_var_overwrite
 
 cmake_minimum_required(VERSION 3.0.2)
-project(seerep_ros_communication VERSION 0.1)
+project(seerep_ros_comm VERSION 0.1)
 
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
