+++ seerep_ros/seerep_ros_comm/include/seerep_ros_comm/clientImagesWithDetection.h
