+++ seerep_ros/seerep_ros_comm/launch/clientImagesWithDetection.launch
